# service/swagger/ducklake-api.yaml
openapi: 3.0.3
info:
  title: DuckLake Data Lakehouse API
  description: |
    REST API for DuckLake - Git-like branching for data lakehouses.

    ## Overview
    This API provides endpoints to:
    - Execute SQL queries and return tabular results
    - Manage data branches (create, delete, search, switch)
    - Query branch metadata and statistics
    - Perform time-travel queries
    - Connect to multiple catalog types (DuckLake, Databricks, Snowflake)

    ## Catalog Types
    The service supports multiple catalog types:
    - **DUCKLAKE**: Native DuckLake catalogs with full read/write access and branching support
    - **DATABRICKS**: Databricks Unity Catalog (read-only, uses OAuth service principal)
    - **SNOWFLAKE**: Snowflake database (read-only, not yet implemented)

    ## Authentication
    All endpoints require Bearer token authentication.

    ## Error Handling
    Errors follow Google's gRPC status code conventions.

  version: 1.0.0
  contact:
    name: DuckLake Team
    email: support@ducklake.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ducklake.io/v1
    description: Production server
  - url: https://api.staging.ducklake.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /query:
    post:
      tags:
        - Query
      summary: Execute SQL query
      description: |
        Executes a SQL query that can span multiple catalogs.

        The service automatically:
        1. Parses the SQL to identify referenced catalogs
        2. Looks up catalog configurations from the registry
        3. Attaches catalogs as needed (using connection pool)
        4. Applies branch context per catalog
        5. Executes the query and returns results

        **Cross-catalog join example:**
        ```sql
        SELECT t.*, p.price
        FROM xva_desk.trades.positions t
        JOIN market_db.main.prices p ON t.instrument_id = p.id
        ```
      operationId: executeQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # CATALOG REGISTRY
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
...
  schemas:
...
    QueryRequest:
      type: object
      required:
        - sql
      properties:
        sql:
          type: string
          description: |
            SQL query to execute. Can reference multiple catalogs.
            Use `catalog.schema.table` format for cross-catalog queries.
          example: "SELECT t.*, p.price FROM xva_desk.trades.positions t JOIN market_db.main.prices p ON t.instrument_id = p.id"
        branch_context:
          type: object
          additionalProperties:
            type: string
          description: |
            Branch context per catalog. Keys are catalog IDs, values are branch names.
            If not specified for a catalog, uses 'main' branch.
          example:
            xva_desk: "feature_analytics"
            market_db: "main"
        snapshot_context:
          type: object
          additionalProperties:
            type: integer
          description: |
            Snapshot context per catalog for time-travel queries.
            Keys are catalog IDs, values are snapshot IDs.
          example:
            xva_desk: 34
        parameters:
          type: object
          additionalProperties: true
          description: Query parameters for parameterized queries
          example:
            start_date: "2024-01-01"
            end_date: "2024-12-31"
        timeout_ms:
          type: integer
          format: int64
          description: Query timeout in milliseconds (default 30000)
          example: 60000
          default: 30000
    QueryResponse:
      type: object
      properties:
        columns:
          type: array
          description: Column metadata
          items:
            $ref: '#/components/schemas/ColumnInfo'
        rows:
          type: array
          description: Result rows
          items:
            type: array
            items:
              oneOf:
                - type: string
                - type: number
                - type: boolean
                - type: 'null'
        row_count:
          type: integer
          format: int64
          description: Number of rows returned
          example: 10
        execution_time_ms:
          type: number
          format: double
          description: Query execution time in milliseconds
          example: 45.2
        branch:
          type: string
          description: Branch context used for the query
          example: "main"
        snapshot_id:
          type: integer
          format: int64
          description: Snapshot ID at time of query
          example: 34
    ColumnInfo:
      type: object
      properties:
        name:
          type: string
          description: Column name
          example: "customer_id"
        type:
          type: string
          description: Column data type
          example: "BIGINT"
        nullable:
          type: boolean
          description: Whether column allows NULL values
          example: true

    # Branch
...
    Status:
      type: object
      description: |
        Error response following google.rpc.Status format.
        See: https://cloud.google.com/apis/design/errors
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: gRPC status code
          example: 5
        message:
          type: string
          description: Human-readable error message
          example: "Branch not found"
        details:
          type: array
          description: Structured error details
          items:
            oneOf:
              - $ref: '#/components/schemas/BadRequestDetail'
              - $ref: '#/components/schemas/ResourceInfo'
              - $ref: '#/components/schemas/ErrorInfo'
    BadRequestDetail:
      type: object
      description: Describes violations in a client request
      properties:
        '@type':
          type: string
          example: "type.googleapis.com/google.rpc.BadRequest"
        field_violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
    FieldViolation:
      type: object
      properties:
        field:
          type: string
          description: Field path that caused the error
          example: "branch_name"
        description:
          type: string
          description: Description of the violation
          example: "Branch name is required"
    ResourceInfo:
      type: object
      description: Information about the resource involved in the error
      properties:
        '@type':
          type: string
          example: "type.googleapis.com/google.rpc.ResourceInfo"
        resource_type:
          type: string
          description: Type of resource
          example: "Branch"
        resource_name:
          type: string
          description: Name of the resource
          example: "feature_nonexistent"
        owner:
          type: string
          description: Service that owns the resource
          example: "ducklake-service"
        description:
          type: string
          description: Human-readable description
          example: "The requested branch does not exist"
    ErrorInfo:
      type: object
      description: Additional error information
      properties:
        '@type':
          type: string
          example: "type.googleapis.com/google.rpc.ErrorInfo"
        reason:
          type: string
          description: Error reason code
          example: "BRANCH_NOT_FOUND"
        domain:
          type: string
          description: Error domain
          example: "ducklake.io"
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Additional metadata
          example:
            catalog_id: "my_lakehouse"
            branch_name: "feature_nonexistent"

  # --------------------------------------------------------------------------
  # RESPONSES
  # --------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request parameters (gRPC code 3 - INVALID_ARGUMENT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            validationFailed:
              summary: Validation failed
              value:
                code: 3
                message: "Validation failed"
                details:
                  - '@type': "type.googleapis.com/google.rpc.BadRequest"
                    field_violations:
                      - field: "branch_name"
                        description: "Branch name is required"
                      - field: "branch_name"
                        description: "Branch name must start with a letter"
            invalidSQL:
              summary: Invalid SQL syntax
              value:
                code: 3
                message: "SQL syntax error"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "INVALID_SQL"
                    domain: "ducklake.io"
                    metadata:
                      position: "15"
                      expected: "FROM"
    Unauthorized:
      description: Missing or invalid authentication (gRPC code 16 - UNAUTHENTICATED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            unauthenticated:
              summary: No token provided
              value:
                code: 16
                message: "Authentication required"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "CREDENTIALS_MISSING"
                    domain: "ducklake.io"
            tokenExpired:
              summary: Token expired
              value:
                code: 16
                message: "Token has expired"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "TOKEN_EXPIRED"
                    domain: "ducklake.io"
...
    NotFound:
      description: Resource not found (gRPC code 5 - NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            catalogNotFound:
              summary: Catalog not found
              value:
                code: 5
                message: "Catalog not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Catalog"
                    resource_name: "nonexistent_catalog"
                    owner: "ducklake-service"
                    description: "The requested catalog does not exist"
            branchNotFound:
              summary: Branch not found
              value:
                code: 5
                message: "Branch not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Branch"
                    resource_name: "feature_nonexistent"
                    owner: "ducklake-service"
                    description: "The requested branch does not exist"
            tableNotFound:
              summary: Table not found
              value:
                code: 5
                message: "Table not found"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ResourceInfo"
                    resource_type: "Table"
                    resource_name: "nonexistent_table"
                    owner: "ducklake-service"
                    description: "The requested table does not exist"
...
    InternalServerError:
      description: Internal server error (gRPC code 13 - INTERNAL)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          examples:
            internalError:
              summary: Internal error
              value:
                code: 13
                message: "Internal server error"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "INTERNAL_ERROR"
                    domain: "ducklake.io"
            metadataConnectionError:
              summary: Metadata database connection error
              value:
                code: 14
                message: "Unable to connect to metadata database"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "METADATA_CONNECTION_ERROR"
                    domain: "ducklake.io"
                    metadata:
                      retry_after: "30"
            storageError:
              summary: Storage access error
              value:
                code: 13
                message: "Unable to access data storage"
                details:
                  - '@type': "type.googleapis.com/google.rpc.ErrorInfo"
                    reason: "STORAGE_ERROR"
                    domain: "ducklake.io"